Add HTST
This is insted of doing this in the ingrss controller

https://github.com/FlowFuse/CloudProject/issues/286
